# Optimise memory view getters