While working on cloudflare/serde-wasm-bindgen#34, I noticed that a significant amount of time was spent getting memory views. Most of that time was spent in the getter for `WebAssembly.Memory.buffer`, so I've changed the memory view getters to check for invalidation without comparing to the wasm memory, by checking if the length is 0 (which indicates the buffer being detached).

`wasm-bindgen` benchmarks:

| Benchmark                                             | Before      | After       | Change  |
| ----------------------------------------------------- | ------------| ----------- | ------- |
| Access Node.nodeType with final                       | 364/s       | 383/s       | +5.22%  |
| Access Node.nodeType with structural                  | 1,166/s     | 1,189/s     | +1.97%  |
| Call a custom JS class Foo.bar method with final      | 6,248/s     | 6,189/s     | -0.94%  |
| Call a custom JS class Foo.bar method with structural | 6,642/s     | 6,667/s     | +0.38%  |
| Pass ascii_small to/from wasm-bindgen                 | 1,281,297/s | 1,823,742/s | +43.04% |
| Pass ascii_medium to/from wasm-bindgen                | 1,231,065/s | 1,723,498/s | +40.00% |
| Pass ascii_number to/from wasm-bindgen                | 1,161,427/s | 1,591,248/s | +37.01% |
| Pass ascii_date to/from wasm-bindgen                  | 918,950/s   | 1,138,144/s | +23.85% |
| Pass ascii_url to/from wasm-bindgen                   | 899,162/s   | 1,111,610/s | +23.63% |
| Pass ascii_link to/from wasm-bindgen                  | 864,511/s   | 1,105,580/s | +27.89% |
| Pass unicode to/from wasm-bindgen                     | 351,071/s   | 390,481/s   | +11.23% |

`serde-wasm-bindgen` benchmarks (not including cloudflare/serde-wasm-bindgen#34):

| Benchmark              | Before        | After         | Change  |
| ---------------------- | ------------- | ------------- | ------- |
| Parse canada           | 44.01 ms/iter | 34.52 ms/iter | -21.56% |
| Parse citm_catalog     | 10.32 ms/iter | 8.78 ms/iter  | -14.92% |
| Parse twitter          | 8.99 ms/iter  | 8.23 ms/iter  | -8.45%  |
| Serialize canada       | 11.74 ms/iter | 11.59 ms/iter | -1.28%  |
| Serialize citm_catalog | 7.51 ms/iter  | 7.51 ms/iter  | -0.00%  |
| Serialize twitter      | 14.54 ms/iter | 14.07 ms/iter | -3.23%  |